# backend/internal/httpapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitFlame CodePilot Backend",
    "version": "0.2.0",
    "description": "Sprint 2 asynchronous Issue-to-Plan integration with a SERGE-based Agent Engine."
  },
  "servers": [{"url": "http://localhost:8000"}],
  "tags": [
    {"name": "health"},
    {"name": "agent tasks"},
    {"name": "issue workflow"},
    {"name": "recommendations"}
  ],
  "paths": {
    "/health": {"get": {"tags": ["health"], "summary": "Backend health check", "responses": {"200": {"description": "Healthy"}}}},
    "/ready": {"get": {"tags": ["health"], "summary": "Check storage, Redis, and Agent Engine readiness", "responses": {"200": {"description": "All configured dependencies are ready"}, "503": {"description": "At least one dependency is unavailable"}}}},
    "/integrations/gitflame/issues/analyze": {
      "post": {
        "tags": ["issue workflow"],
        "summary": "Validate an issue and enqueue plan generation",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IssueAnalyzeRequest"}}}},
        "responses": {
          "202": {"description": "Agent task queued", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IssueAnalyzeResponse"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "422": {"$ref": "#/components/responses/ValidationError"}
        }
      }
    },
    "/ai/tasks/{taskId}": {
      "get": {
        "tags": ["agent tasks"], "summary": "Get asynchronous generation status",
        "parameters": [{"$ref": "#/components/parameters/TaskID"}],
        "responses": {"200": {"description": "Task state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AgentTask"}}}}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/ai/tasks/{taskId}/retry": {"post": {"tags": ["agent tasks"], "summary": "Retry a task that failed with a temporary Agent Engine error", "parameters": [{"$ref": "#/components/parameters/TaskID"}], "responses": {"202": {"description": "Replacement task queued"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/ValidationError"}}}},
    "/ai/issues/{id}/plan": {"get": {"tags": ["issue workflow"], "summary": "Get the latest generated plan", "parameters": [{"$ref": "#/components/parameters/IssueID"}], "responses": {"200": {"description": "Latest plan"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"description": "Plan is not ready"}}}},
    "/ai/issues/{id}/approve": {"post": {"tags": ["issue workflow"], "summary": "Approve the plan and prepare a generated-files contract", "parameters": [{"$ref": "#/components/parameters/IssueID"}], "responses": {"200": {"description": "Plan approved"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/ValidationError"}}}},
    "/ai/issues/{id}/correct": {
      "post": {
        "tags": ["issue workflow"], "summary": "Enqueue regeneration with feedback and the previous plan", "parameters": [{"$ref": "#/components/parameters/IssueID"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["feedback"], "properties": {"feedback": {"type": "string"}}}}}},
        "responses": {"202": {"description": "Correction task queued"}, "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/ValidationError"}}
      }
    },
    "/ai/issues/{id}/reject": {"post": {"tags": ["issue workflow"], "summary": "Reject a plan", "parameters": [{"$ref": "#/components/parameters/IssueID"}], "responses": {"200": {"description": "Plan rejected"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/ValidationError"}}}},
    "/integrations/gitflame/repositories/{id}/recommendations/analyze": {"post": {"tags": ["recommendations"], "summary": "Analyze repository recommendations", "parameters": [{"$ref": "#/components/parameters/RepositoryID"}], "responses": {"200": {"description": "Recommendation report"}, "400": {"$ref": "#/components/responses/BadRequest"}, "422": {"$ref": "#/components/responses/ValidationError"}}}},
    "/repositories/{id}/recommendations/status": {"get": {"tags": ["recommendations"], "summary": "Get recommendation counters", "parameters": [{"$ref": "#/components/parameters/RepositoryID"}], "responses": {"200": {"description": "Counters"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/repositories/{id}/recommendations/summary": {"get": {"tags": ["recommendations"], "summary": "Get recommendation summary", "parameters": [{"$ref": "#/components/parameters/RepositoryID"}], "responses": {"200": {"description": "Summary"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/repositories/{id}/recommendations": {"get": {"tags": ["recommendations"], "summary": "List recommendations", "parameters": [{"$ref": "#/components/parameters/RepositoryID"}], "responses": {"200": {"description": "Recommendations"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/recommendations/{id}/close": {"patch": {"tags": ["recommendations"], "summary": "Close recommendation", "parameters": [{"$ref": "#/components/parameters/RecommendationID"}], "responses": {"200": {"description": "Closed"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/recommendations/{id}": {"delete": {"tags": ["recommendations"], "summary": "Delete recommendation", "parameters": [{"$ref": "#/components/parameters/RecommendationID"}], "responses": {"204": {"description": "Deleted"}, "404": {"$ref": "#/components/responses/NotFound"}}}}
  },
  "components": {
    "parameters": {
      "TaskID": {"name": "taskId", "in": "path", "required": true, "schema": {"type": "string"}},
      "IssueID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
      "RepositoryID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
      "RecommendationID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
    },
    "responses": {
      "BadRequest": {"description": "Malformed JSON", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
      "NotFound": {"description": "Resource not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
      "ValidationError": {"description": "Invalid input or workflow state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
      "Repository": {"type": "object", "required": ["id", "default_branch"], "properties": {"id": {"type": "string"}, "name": {"type": "string"}, "default_branch": {"type": "string"}, "commit_sha": {"type": "string"}, "web_url": {"type": "string"}}},
      "Issue": {"type": "object", "required": ["id", "title", "body", "author"], "properties": {"id": {"type": "string"}, "title": {"type": "string"}, "body": {"type": "string"}, "author": {"type": "string"}}},
      "RepositoryFile": {"type": "object", "required": ["path", "content"], "properties": {"path": {"type": "string"}, "content": {"type": "string"}}},
      "IssueAnalyzeRequest": {"type": "object", "required": ["repository", "issue", "yaml_config", "repository_files"], "properties": {"repository": {"$ref": "#/components/schemas/Repository"}, "issue": {"$ref": "#/components/schemas/Issue"}, "yaml_config": {"type": "string"}, "repository_files": {"type": "array", "minItems": 1, "items": {"$ref": "#/components/schemas/RepositoryFile"}}, "repository_context": {"type": "array", "deprecated": true, "items": {"type": "string"}}, "metadata": {"type": "object", "additionalProperties": true}}},
      "IssueAnalyzeResponse": {"type": "object", "required": ["session_id", "task_id", "status", "status_url"], "properties": {"session_id": {"type": "string"}, "task_id": {"type": "string"}, "issue_id": {"type": "string"}, "repository_id": {"type": "string"}, "status": {"type": "string", "enum": ["queued"]}, "status_url": {"type": "string"}}},
      "TaskError": {"type": "object", "properties": {"http_status": {"type": "integer", "enum": [400,404,422,502,503,504]}, "code": {"type": "string"}, "detail": {"type": "string"}}},
      "AgentTask": {"type": "object", "required": ["task_id", "session_id", "issue_id", "type", "status", "attempt"], "properties": {"task_id": {"type": "string"}, "session_id": {"type": "string"}, "issue_id": {"type": "string"}, "type": {"type": "string", "enum": ["generate", "correction"]}, "status": {"type": "string", "enum": ["queued", "processing", "completed", "failed"]}, "attempt": {"type": "integer", "minimum": 1}, "plan_markdown": {"type": "string"}, "tool_execution_summary": {"type": "string"}, "relevant_files": {"type": "array", "items": {"type": "object", "properties": {"path": {"type": "string"}, "reason": {"type": "string"}}}}, "model": {"type": "string"}, "usage": {"type": "object", "properties": {"prompt_tokens": {"type": "integer"}, "completion_tokens": {"type": "integer"}, "tool_calls": {"type": "integer"}}}, "error": {"$ref": "#/components/schemas/TaskError"}}},
      "Problem": {"type": "object", "required": ["status", "code", "detail"], "properties": {"status": {"type": "integer"}, "code": {"type": "string"}, "detail": {"type": "string"}}}
    }
  }
}